fix(app): move markdown parsing to worker
