+++ packages/session-ui/src/components/markdown-cache.tsx
@@ -1,6 +1,6 @@
 import { checksum } from "@opencode-ai/core/util/encode"
 import DOMPurify from "dompurify"
-import { project } from "./markdown-stream"
+import { parseMarkdown } from "./markdown-worker"
 
 export type MarkdownCacheEntry = {
   raw: string
@@ -52,27 +52,18 @@ export function touchCachedMarkdown(key: string, value: MarkdownCacheEntry) {
   cache.delete(first)
 }
 
-export async function preloadMarkdown(
-  text: string,
-  cacheKey: string,
-  parser: { parse(text: string): string | Promise<string> },
-) {
-  await Promise.all(
-    project(undefined, text, false).blocks.map(async (block, index) => {
-      if (block.mode === "code") return
-      const key = `${cacheKey}:${index}:${block.mode}`
-      const cached = getCachedMarkdown(key)
-      if (cached?.raw === block.raw) {
-        touchCachedMarkdown(key, cached)
-        return
-      }
-      const hash = checksum(block.raw)
-      if (!hash) return
-      touchCachedMarkdown(key, {
-        raw: block.raw,
-        hash,
-        html: sanitizeMarkdown(await Promise.resolve(parser.parse(block.src))),
-      })
-    }),
-  )
+export async function preloadMarkdown(text: string, cacheKey: string) {
+  const key = `${cacheKey}:0:full`
+  const cached = getCachedMarkdown(key)
+  if (cached?.raw === text) {
+    touchCachedMarkdown(key, cached)
+    return
+  }
+  const hash = checksum(text)
+  if (!hash) return
+  touchCachedMarkdown(key, {
+    raw: text,
+    hash,
+    html: sanitizeMarkdown(await parseMarkdown(text)),
+  })
 }
