+++ packages/session-ui/src/components/markdown-projection.ts
@@ -1,0 +1,7 @@
+import type { Block } from "./markdown-stream"
+
+export function canReusePendingBlock(current: Pick<Block, "mode" | "raw"> | undefined, next: Block) {
+  if (!current || current.mode !== next.mode) return false
+  if (next.mode === "code") return next.raw.startsWith(current.raw)
+  return current.raw === next.raw
+}
