+++ packages/session-ui/src/components/markdown-worker-protocol.ts
@@ -1,27 +1,32 @@
-import type { ThemeRegistrationResolved } from "shiki"
+import type { Projection } from "./markdown-stream"
 
 export type MarkdownToken = [content: string, style: string]
 
 export type MarkdownWorkerRequest =
-  | { type: "init"; theme: ThemeRegistrationResolved }
+  | { type: "parse"; id: number; text: string }
+  | { type: "project"; id: number; key: string; text: string; live: boolean }
   | { type: "highlight"; id: number; key: string; text: string; language: string; complete?: boolean }
   | { type: "dispose"; key: string }
 
 export type MarkdownWorkerResponse =
+  | { type: "parse"; id: number; html: string }
+  | { type: "project"; id: number; key: string; projection: Projection }
   | {
       type: "highlight"
       id: number
       key: string
+      language: string
       reset: boolean
       stable: MarkdownToken[]
       unstable: MarkdownToken[]
     }
-  | { type: "error"; id: number; key: string; message: string }
+  | { type: "error"; id: number; key?: string; message: string }
   | { type: "superseded"; id: number; key: string }
 
 export type MarkdownWorkerState = {
   id: number
   generation: number
+  language: string
   stable: MarkdownToken[]
   unstable: MarkdownToken[]
 }
@@ -42,6 +47,7 @@ export function applyMarkdownWorkerResponse(
   return {
     id: response.id,
     generation: (state?.generation ?? 0) + (response.reset ? 1 : 0),
+    language: response.language,
     stable: response.reset ? response.stable : [...(state?.stable ?? []), ...response.stable],
     unstable: response.unstable,
   }
