+++ packages/ui/package.json
@@ -94,7 +94,6 @@
     "katex": "0.16.27",
     "luxon": "catalog:",
     "marked": "catalog:",
-    "marked-katex-extension": "5.1.6",
     "marked-shiki": "catalog:",
     "morphdom": "2.7.8",
     "motion": "12.34.5",
