+++ packages/ui/src/context/marked-regression.test.ts
@@ -1,9 +1,8 @@
 import { expect, test } from "bun:test"
 import { Marked } from "marked"
-import { markedCodeSpanBoundary } from "./marked-code-span"
 
 test("preserves code spans adjacent to tildes", async () => {
-  const marked = new Marked(markedCodeSpanBoundary)
+  const marked = new Marked()
 
   expect(await marked.parse("~`0.1576` to measurement-window-only `0.00092`")).toBe(
     "<p>~<code>0.1576</code> to measurement-window-only <code>0.00092</code></p>\n",
